Finishing up merge from 5-general-refactor.
