+++ src/main/java/frc/robot/chassis/TankSubsystem.java
@@ -4,7 +4,7 @@ package frc.robot.chassis;
 import com.ctre.phoenix.motorcontrol.TalonSRXControlMode;
 import com.ctre.phoenix.motorcontrol.can.TalonSRX;
 
-import com.ctre.phoenix6.hardware.CANcoder;
+import edu.wpi.first.wpilibj.Encoder;
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
 import frc.robot.RobotMap;
 
@@ -20,16 +20,17 @@ public class TankSubsystem extends SubsystemBase{
 
     private final TalonSRX leftmotor=new TalonSRX(RobotMap.CHASSIS_LEFT_MOTOR);
     private final TalonSRX leftmotorfollow=new TalonSRX(RobotMap.CHASSIS_LEFT_FOLLOW_MOTOR);
-    private final TalonSRX rightmotor=new TalonSRX(RobotMap.CHASSIS_RIGHT_ENCODER);
+    private final TalonSRX rightmotor=new TalonSRX(RobotMap.CHASSIS_RIGHT_MOTOR);
     private final TalonSRX rightmotorfollow=new TalonSRX(RobotMap.CHASSIS_RIGHT_FOLLOW_MOTOR);// can ids of the talons
-    private final CANcoder leftEncoder = new CANcoder(RobotMap.CHASSIS_LEFT_ENCODER);
-    private final CANcoder rightEncoder = new CANcoder(RobotMap.CHASSIS_RIGHT_ENCODER);
+    private final Encoder leftEncoder = new Encoder(RobotMap.CHASSIS_LEFT_ENCODER[0], RobotMap.CHASSIS_LEFT_ENCODER[1]);
+    private final Encoder rightEncoder = new Encoder(RobotMap.CHASSIS_RIGHT_ENCODER[0], RobotMap.CHASSIS_RIGHT_ENCODER[1]);
 
     public void init(){
         rightmotor.setInverted(true);
         rightmotorfollow.setInverted(true);
         rightmotorfollow.follow(rightmotor);//inverting properly(change if robot moves backwords)
         leftmotorfollow.follow(leftmotor);
+        zeroPosition();
     }
 
     public void drive(double leftStick, double rightStick){
@@ -37,12 +38,17 @@ public class TankSubsystem extends SubsystemBase{
         rightmotor.set(TalonSRXControlMode.PercentOutput,rightStick);
     }
 
+    public void zeroPosition() {
+        leftEncoder.reset();
+        rightEncoder.reset();
+    }
+
     public double getLeftDrivePosition(){
-        return leftEncoder.getPosition().getValueAsDouble();
+        return leftEncoder.getDistance();
     }
 
     public double getRightDrivePosition(){
-        return rightEncoder.getPosition().getValueAsDouble();
+        return rightEncoder.getDistance();
     }
 
     public double[] getDrivePosition() {
